Remove old municipal and region hooks, made sortFeatures generic to strictly type feature properties
