+++ src/utils/useMunicipalityData.ts
@@ -9,15 +9,6 @@ export type TMunicipalityMetricName = keyof Pick<
 
 export default useMunicipalityData;
 
-function createMunicipalCodeFilter<T extends { gmcode: string }>(
-  municipalCodes?: string[]
-) {
-  if (municipalCodes) {
-    return (item: T): boolean => municipalCodes.indexOf(item.gmcode) > -1;
-  }
-  return (_item: T): boolean => true;
-}
-
 /**
  * This hook return a list of municipality data where each item has a value property
  * that is set based on the given metric name.
@@ -30,28 +21,30 @@ function createMunicipalCodeFilter<T extends { gmcode: string }>(
 function useMunicipalityData<
   T extends TMunicipalityMetricName,
   K extends Municipalities[T]
->(
-  metricName?: T,
-  municipalCodes?: string[]
-): (K[number] & { value: number })[] {
+>(metricName?: T): Record<string, K[number] & { value: number }> {
   const { data } = useSWR<Municipalities>('/json/municipalities.json');
 
   const metricItems = metricName ? data?.[metricName] : undefined;
 
   return useMemo(() => {
     if (!metricItems) {
-      return [];
+      return {};
     }
 
-    const filteredData = (metricItems as any[]).map<
-      K[number] & { value: number }
-    >((item: K[number]): K[number] & { value: number } => ({
-      ...item,
-      value: (item as any)[metricName],
-    }));
-
-    return filteredData.filter(
-      createMunicipalCodeFilter<K[number] & { value: number }>(municipalCodes)
+    const filteredData = (metricItems as any[]).reduce(
+      (
+        aggr,
+        item: K[number]
+      ): Record<string, K[number] & { value: number }> => {
+        aggr[item.gmcode] = {
+          ...item,
+          value: (item as any)[metricName],
+        };
+        return aggr;
+      },
+      {}
     );
-  }, [metricItems, metricName, municipalCodes]);
+
+    return filteredData;
+  }, [metricItems, metricName]);
 }
